ISSUE #2243 - move rule func to helper file
